add plan

+++ src/main/java/demo/model/ViajePlanPK.java
@@ -12,26 +12,26 @@ public class ViajePlanPK implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Column(name = "IdViaje")
-    private int IdViaje;
+    private Long IdViaje;
 
     @Column(name = "IdPlan")
-    private int IdPlan;
+    private Long IdPlan;
 
 	public ViajePlanPK(){
 		
 	}
     
-	public ViajePlanPK(int IdViaje, int IdPlan) {
+	public ViajePlanPK(Long IdViaje, Long IdPlan) {
 		super();
 		this.IdViaje = IdViaje;
 		this.IdPlan = IdPlan;
 	}
 
-	public int getId_Estudiante() {
+	public Long getId_Estudiante() {
 		return IdViaje;
 	}
 
-	public int getId_Carrera() {
+	public Long getId_Carrera() {
 		return IdPlan;
 	}
